pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "garmin-mcp"
version = "0.1.0"
description = "MCP server to access Garmin data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv==1.2.1",
    "python-dateutil==2.9.0",
    "garminconnect==0.2.38",
    "requests==2.32.3",
    "mcp==1.3.0",
    "garth>=0.5.17,<0.6.0",
]

[project.scripts]
garmin-mcp = "garmin_mcp:main"
garmin-mcp-auth = "garmin_mcp.auth_cli:main"

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.2",
    "pytest-mock>=3.14.0",
    "pytest-timeout>=2.3.1",
]

[tool.uv.sources]
garmin-mcp = { workspace = true }
